+++ src/main/java/net/geforcemods/securitycraft/blockentities/SonicSecuritySystemBlockEntity.java
@@ -68,6 +68,7 @@ public class SonicSecuritySystemBlockEntity extends CustomizableBlockEntity impl
 	private int listenPos = 0;
 	private boolean tracked = false;
 	private boolean shutDown = false;
+	private boolean disableBlocksWhenTuneIsPlayed = false;
 
 	public SonicSecuritySystemBlockEntity() {
 		super(SCContent.SONIC_SECURITY_SYSTEM_BLOCK_ENTITY.get());
@@ -198,6 +199,7 @@ public class SonicSecuritySystemBlockEntity extends CustomizableBlockEntity impl
 		tag.putBoolean("correctTuneWasPlayed", correctTuneWasPlayed);
 		tag.putInt("powerCooldown", powerCooldown);
 		tag.putBoolean("shutDown", shutDown);
+		tag.putBoolean("disableBlocksWhenTuneIsPlayed", disableBlocksWhenTuneIsPlayed);
 		return tag;
 	}
 
@@ -232,6 +234,7 @@ public class SonicSecuritySystemBlockEntity extends CustomizableBlockEntity impl
 		correctTuneWasPlayed = tag.getBoolean("correctTuneWasPlayed");
 		powerCooldown = tag.getInt("powerCooldown");
 		shutDown = tag.getBoolean("shutDown");
+		disableBlocksWhenTuneIsPlayed = tag.getBoolean("disableBlocksWhenTuneIsPlayed");
 	}
 
 	/**
@@ -496,6 +499,14 @@ public class SonicSecuritySystemBlockEntity extends CustomizableBlockEntity impl
 	@Override
 	public void setShutDown(boolean shutDown) {
 		this.shutDown = shutDown;
+	}
+
+	public void setDisableBlocksWhenTuneIsPlayed(boolean disableBlocksWhenTuneIsPlayed) {
+		this.disableBlocksWhenTuneIsPlayed = disableBlocksWhenTuneIsPlayed;
+	}
+
+	public boolean disablesBlocksWhenTuneIsPlayed() {
+		return disableBlocksWhenTuneIsPlayed;
 	}
 
 	@Override