+++ src/main/java/net/geforcemods/securitycraft/network/server/SyncSSSSettingsOnServer.java
@@ -76,6 +76,9 @@ public class SyncSSSSettingsOnServer {
 					break;
 				case REMOVE_POS:
 					sss.delink(posToRemove, false);
+					break;
+				case INVERT_FUNCTIONALITY:
+					sss.setDisableBlocksWhenTuneIsPlayed(!sss.disablesBlocksWhenTuneIsPlayed());
 			}
 		}
 	}
@@ -88,6 +91,7 @@ public class SyncSSSSettingsOnServer {
 		RECORDING_ON,
 		RECORDING_OFF,
 		CLEAR_NOTES,
-		REMOVE_POS;
+		REMOVE_POS,
+		INVERT_FUNCTIONALITY
 	}
 }